+++ cmake/unit_testing.cmake
@@ -13,7 +13,7 @@ enable_testing()
 add_subdirectory( src/ACEtk/ContinuousEnergyTable/test )
 add_subdirectory( src/ACEtk/DosimetryTable/test )
 add_subdirectory( src/ACEtk/ThermalScatteringTable/test )
-add_subdirectory( src/ACEtk/PhotoAtomicTable/test )
+add_subdirectory( src/ACEtk/PhotoatomicTable/test )
 add_subdirectory( src/ACEtk/Table/Data/Parse/test )
 add_subdirectory( src/ACEtk/Table/Data/test )
 add_subdirectory( src/ACEtk/Table/Header/test )
@@ -78,14 +78,14 @@ add_subdirectory( src/ACEtk/block/ThermalScatteringDiscreteCosines/test )
 add_subdirectory( src/ACEtk/block/ThermalScatteringDiscreteCosinesWithProbability/test )
 add_subdirectory( src/ACEtk/block/ThermalScatteringElasticAngularDistributionBlock/test )
 add_subdirectory( src/ACEtk/block/ThermalScatteringInelasticAngularDistributionBlock/test )
-add_subdirectory( src/ACEtk/block/PhotoAtomicPrincipalCrossSectionBlock/test )
+add_subdirectory( src/ACEtk/block/PhotoatomicPrincipalCrossSectionBlock/test )
 add_subdirectory( src/ACEtk/block/IncoherentScatteringFunctionBlock/test )
 add_subdirectory( src/ACEtk/block/CoherentFormFactorBlock/test )
-add_subdirectory( src/ACEtk/block/PhotoAtomicFluorescenceDataBlock/test )
-add_subdirectory( src/ACEtk/block/PhotoAtomicHeatingNumbersBlock/test )
-add_subdirectory( src/ACEtk/block/PhotoAtomicComptonProfile/test )
-add_subdirectory( src/ACEtk/block/PhotoAtomicComptonProfileBlock/test )
-add_subdirectory( src/ACEtk/block/PhotoAtomicElectronShellBlock/test )
+add_subdirectory( src/ACEtk/block/PhotoatomicFluorescenceDataBlock/test )
+add_subdirectory( src/ACEtk/block/PhotoatomicHeatingNumbersBlock/test )
+add_subdirectory( src/ACEtk/block/PhotoatomicComptonProfile/test )
+add_subdirectory( src/ACEtk/block/PhotoatomicComptonProfileBlock/test )
+add_subdirectory( src/ACEtk/block/PhotoatomicElectronShellBlock/test )
 add_subdirectory( src/ACEtk/details/verify/CDF/test )
 add_subdirectory( src/ACEtk/details/verify/Positive/test )
 add_subdirectory( src/ACEtk/details/verify/Sorted/test )
