--- src/ACEtk/block/PhotoAtomicComptonProfile/test/CMakeLists.txt
@@ -1,18 +1,0 @@
-
-add_executable( ACEtk.block.PhotoAtomicComptonProfile.test PhotoAtomicComptonProfile.test.cpp )
-target_compile_options( ACEtk.block.PhotoAtomicComptonProfile.test PRIVATE ${${PREFIX}_common_flags}
-$<$<BOOL:${strict}>:${${PREFIX}_strict_flags}>$<$<CONFIG:DEBUG>:
-${${PREFIX}_DEBUG_flags}
-$<$<BOOL:${coverage}>:${${PREFIX}_coverage_flags}>>
-$<$<CONFIG:RELEASE>:
-${${PREFIX}_RELEASE_flags}
-$<$<BOOL:${link_time_optimization}>:${${PREFIX}_link_time_optimization_flags}>
-$<$<BOOL:${nonportable_optimization}>:${${PREFIX}_nonportable_optimization_flags}>>
-
-${CXX_appended_flags} ${ACEtk_appended_flags} )
-target_link_libraries( ACEtk.block.PhotoAtomicComptonProfile.test PUBLIC ACEtk )
-file( GLOB resources "resources/*" )
-foreach( resource ${resources})
-    file( COPY "${resource}" DESTINATION "${CMAKE_CURRENT_BINARY_DIR}" )
-endforeach()
-add_test( NAME ACEtk.block.PhotoAtomicComptonProfile COMMAND ACEtk.block.PhotoAtomicComptonProfile.test )
